requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0.3
chromadb>=0.4.10
openai>=1.3.0
pydantic>=2.3.0
sentence-transformers>=2.2.2
pyyaml>=6.0.1
watchdog>=3.0.0
google-genai>=0.1.0
